refactor: user module refactor
