+++ src/follow-project/service/follow-project.service.ts
@@ -4,7 +4,7 @@ import { Data } from 'src/follow-project/types/follow';
 import { i18n } from 'src/i18n';
 import { LoggerService } from 'src/logger/logger.service';
 import { ProjectsService } from 'src/projects/services/projects/projects.service';
-import { UserService } from 'src/user/services/user/user.service';
+import { UserService } from 'src/user/services/user.service';
 
 @Injectable()
 export class FollowProjectService {
